# docs/architecture/api/openapi.yaml
openapi: 3.1.0
info:
  title: FieldLog API
  version: 0.1.0
  description: >
    HTTP API the offline-first FieldLog client syncs against, and that the future
    contractor mobile app and manager web app consume. Component schemas are
    structurally derived from the client Zod source of truth in
    `src/domain/schemas.ts`; server-only fields (serverId == id, etag,
    createdByUserId) are additive and called out. The system captures immutable
    pesticide-application evidence; nothing may alter a record after `locked`
    except appending an `exported` event.
  contact:
    name: FieldLog (ComplianceLog)
    url: https://github.com/toddAston/ComplianceLog
servers:
  - url: https://api.fieldlog.example/v1
    description: Production (placeholder)
  - url: http://localhost:8080/v1
    description: Local compose
...
security:
  - bearerAuth: []
paths:
...
  /application-records:
    get:
      tags: [records]
      summary: List application records
      description: >
        Applicators see only records they created; managers see all records in
        their organization. Use `workflowStatus=pending_review` for the review queue.
      operationId: listApplicationRecords
      parameters:
        - name: workflowStatus
          in: query
          required: false
          schema: { $ref: "#/components/schemas/WorkflowStatus" }
        - name: syncStatus
          in: query
          required: false
          schema: { $ref: "#/components/schemas/SyncStatus" }
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [systemCreatedAt, -systemCreatedAt, applicationDate, -applicationDate]
            default: -systemCreatedAt
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200":
          description: Page of application records.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationRecordPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [records]
      summary: Create a draft application record
      description: >
        Accepts a client-generated UUID. Server forces workflowStatus=draft and
        re-validates contractorInputs against the Zod schema.
      operationId: createApplicationRecord
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationRecordCreate" }
      responses:
        "201":
          description: Draft created.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationRecord" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /application-records/{recordId}/submit:
    parameters:
      - { $ref: "#/components/parameters/RecordId" }
    post:
      tags: [records]
      summary: Submit a draft (draft → pending_review)
      description: >
        Server re-runs compliance checks; a `blocked`-severity failure (e.g.
        RUP_UNCERTIFIED) rejects with 422. On success it freezes the Product
        Snapshot in the same transaction and appends submitted +
        product_snapshot_created + compliance_check_run events.
      operationId: submitApplicationRecord
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
        - { $ref: "#/components/parameters/IfMatch" }
      responses:
        "200":
          description: Submitted; now pending_review.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationRecord" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token from /auth/login or /auth/refresh.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
      description: Opaque cursor from a prior page's `nextCursor`.
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: Client-generated UUID; the server replays the stored result on retry.
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string }
      description: Current ETag of the resource for optimistic concurrency.
...
    RecordId:
      name: recordId
      in: path
      required: true
      schema: { type: string, format: uuid }
  headers:
...
    ETag:
      description: Opaque concurrency token for the resource.
      schema: { type: string }
  responses:
    Unauthorized:
      description: Authentication required or invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: AUTH_REQUIRED, message: Authentication required., requestId: 01J9Z3K8F2 }
    Forbidden:
      description: Authenticated but not permitted (role or ownership).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: FORBIDDEN, message: You do not have access to this resource., requestId: 01J9Z3K8F2 }
    NotFound:
      description: Resource not found (or not in the caller's org).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: NOT_FOUND, message: Resource not found., requestId: 01J9Z3K8F2 }
    Conflict:
      description: Conflicting state (e.g. duplicate, locked record, invalid transition).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: STATUS_TRANSITION_INVALID, message: Only draft records can be submitted., requestId: 01J9Z3K8F2 }
    PreconditionFailed:
      description: If-Match ETag did not match current state (stale client).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: CONFLICT_STALE_RECORD, message: Record was modified; refetch and retry., requestId: 01J9Z3K8F2 }
    ValidationError:
      description: Malformed request (bad JSON, missing required header).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: VALIDATION_FAILED, message: Request body failed validation., requestId: 01J9Z3K8F2 }
    UnprocessableEntity:
      description: Well-formed but semantically invalid (Zod/domain rule failure).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: VALIDATION_FAILED
              message: Restricted-use product applied by uncertified applicator — blocked by RSMo 281.037(9).
              requestId: 01J9Z3K8F2
...
    InternalError:
      description: Unexpected server error. Body never leaks internals.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: INTERNAL, message: An unexpected error occurred., requestId: 01J9Z3K8F2 }
  schemas:
...
    RupStatus:
      type: string
      enum: [yes, no, unknown]
    WorkflowStatus:
      type: string
      enum: [draft, submitted, pending_review, needs_correction, accepted, locked, exported]
    SyncStatus:
      type: string
      enum: [local_only, queued, syncing, synced, sync_failed]
    ReviewStatus:
      type: string
      enum: [not_reviewed, accepted, needs_correction, rejected]
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { type: string, description: Stable machine code (see error_codes.md). }
            message: { type: string, description: Safe, human-readable summary. Never a stack trace. }
            requestId: { type: string, description: Correlates to server logs. }
            details:
              type: array
              description: Optional field-level validation issues.
              items:
                type: object
                additionalProperties: false
                properties:
                  path: { type: string }
                  message: { type: string }

    # ---- Health ----
...
    WeatherSnapshot:
      type: object
      additionalProperties: false
      required: [source, capturedAt]
      properties:
        source: { type: string, enum: [nws_observation, nws_forecast_grid, manual, stale_cache] }
        stationId: { type: string }
        observedAt: { type: string, format: date-time }
        capturedAt: { type: string, format: date-time }
    ProductSnapshot:
      type: object
      additionalProperties: false
      required: [id, applicationRecordId, productName, epaRegistrationNumber, rupStatus, catalogVersion, snapshotCreatedAt]
      properties:
        id: { type: string, format: uuid }
        applicationRecordId: { type: string, format: uuid }
        sourceProductId: { type: string, format: uuid }
        productName: { type: string }
        epaRegistrationNumber: { type: string, pattern: "^[A-Za-z0-9-]+$" }
        rupStatus: { $ref: "#/components/schemas/RupStatus" }
        catalogVersion: { type: string }
        snapshotCreatedAt: { type: string, format: date-time }
    ContractorInputs:
      type: object
      additionalProperties: false
      required:
        - applicatorId
        - applicatorName
        - company
        - farmId
        - farmName
        - fieldId
        - fieldName
        - cropOrSite
        - acresTreated
        - productName
        - epaRegistrationNumber
        - rupStatus
        - applicationDate
        - startTime
        - applicationMethod
        - rateApplied
        - totalAmountApplied
        - temperature
        - windSpeed
        - windDirection
        - attestationConfirmed
      properties:
        applicatorId: { type: string, format: uuid }
        applicatorName: { type: string }
        company: { type: string }
        certificationNumber: { type: string }
        farmId: { type: string, format: uuid }
        farmName: { type: string }
        fieldId: { type: string, format: uuid }
        fieldName: { type: string }
        cropOrSite: { type: string }
        acresTreated: { type: string, description: "Wire shape is a string (client); coerced to numeric server-side." }
        productId: { type: string, format: uuid }
        productName: { type: string }
        epaRegistrationNumber: { type: string, pattern: "^[A-Za-z0-9-]+$" }
        rupStatus: { $ref: "#/components/schemas/RupStatus" }
        catalogVersion: { type: string }
        applicationDate: { type: string, format: date }
        startTime: { type: string, description: "HH:mm time of day." }
        endTime: { type: string }
        applicationMethod: { type: string }
        rateApplied: { type: string }
        totalAmountApplied: { type: string }
        targetPest: { type: string }
        phi: { type: string }
        temperature: { type: string }
        windSpeed: { type: string }
        windDirection: { type: string }
        weatherNotes: { type: string }
        weatherSnapshot: { $ref: "#/components/schemas/WeatherSnapshot" }
        attestationConfirmed: { type: boolean }
        submittedBy: { type: string }
        submittedAt: { type: string, format: date-time }
...
    ManagerInputs:
      type: object
      additionalProperties: false
      required: [reviewStatus]
      properties:
        reviewStatus: { $ref: "#/components/schemas/ReviewStatus" }
        reviewedBy: { type: string }
        reviewedAt: { type: string, format: date-time }
        reviewNotes: { type: string }
    SystemCapturedFields:
      type: object
      additionalProperties: false
      required: [createdAt, createdOffline, lastUpdatedAt]
      properties:
        createdAt: { type: string, format: date-time }
        createdOffline: { type: boolean }
        lastUpdatedAt: { type: string, format: date-time }
        lockedAt: { type: string, format: date-time }
        catalogVersion: { type: string }
    ApplicationRecord:
      type: object
      additionalProperties: false
      required:
        - id
        - organizationId
        - workflowStatus
        - syncStatus
        - contractorInputs
        - managerInputs
        - system
        - complianceReviewRequired
        - etag
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        workflowStatus: { $ref: "#/components/schemas/WorkflowStatus" }
        syncStatus: { $ref: "#/components/schemas/SyncStatus" }
        contractorInputs: { $ref: "#/components/schemas/ContractorInputs" }
        managerInputs: { $ref: "#/components/schemas/ManagerInputs" }
        system: { $ref: "#/components/schemas/SystemCapturedFields" }
        productSnapshotId: { type: string, format: uuid }
        productSnapshot:
          allOf:
            - { $ref: "#/components/schemas/ProductSnapshot" }
          description: Frozen product reference; embedded once the record is submitted.
        complianceReviewRequired: { type: boolean }
        etag:
          type: string
          description: Server-added concurrency token. Not present on the v0.1 client.
        createdByUserId:
          type: string
          format: uuid
          description: Server-added. Not present on the v0.1 client.
    ApplicationRecordCreate:
      type: object
      additionalProperties: false
      required: [id, contractorInputs]
      description: >
        Draft creation. Client supplies its UUID and contractorInputs; the server
        forces workflowStatus=draft, syncStatus, system fields and managerInputs.
      properties:
        id: { type: string, format: uuid }
        contractorInputs: { $ref: "#/components/schemas/ContractorInputs" }
        createdOffline: { type: boolean, default: true }
    ApplicationRecordPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ApplicationRecord" }
        page: { $ref: "#/components/schemas/PageMeta" }
...
    PageMeta:
      type: object
      additionalProperties: false
      required: [hasMore]
      properties:
        hasMore: { type: boolean }
        nextCursor:
          type: string
          description: Pass as `cursor` to fetch the next page. Absent on the last page.

    # ---- Sync ----
...